Add hashtags (if present) in URL to tags dictionary.
It seems that Sentry strips URLs off their hashtags and for some users this might pose a problem. My suggested solution is to add the hashes to the tags dictionary if they are present in the url (as told by the window.location.hash property). This way they will show up in the Sentry GUI under Tags.

Signed-off-by: parhamfh <parhamfh@kth.se>